Move the prompt into the agent service
