+++ src/steamship/agents/service/agent_service.py
@@ -1,10 +1,14 @@
 import logging
+import uuid
+from typing import List, Optional
 
-from steamship import SteamshipError, Task
+from steamship import Block, SteamshipError, Task
+from steamship.agents.llms import OpenAI
 from steamship.agents.logging import AgentLogging
 from steamship.agents.schema import Action, Agent, FinishAction
-from steamship.agents.schema.context import AgentContext
-from steamship.invocable import PackageService
+from steamship.agents.schema.context import AgentContext, Metadata
+from steamship.agents.utils import with_llm
+from steamship.invocable import PackageService, post
 
 
 class AgentService(PackageService):
@@ -79,3 +83,51 @@ class AgentService(PackageService):
         for func in context.emit_funcs:
             logging.info(f"Emitting via function: {func.__name__}")
             func(action.output, context.metadata)
+
+    @post("prompt")
+    def prompt(self, prompt: str, **kwargs) -> List[Block]:
+        """Run an agent with the provided text as the input."""
+
+        # AgentContexts serve to allow the AgentService to run agents
+        # with appropriate information about the desired tasking.
+        # Here, we create a new context on each prompt, and append the
+        # prompt to the message history stored in the context.
+        context_id = uuid.uuid4()
+        context = AgentContext.get_or_create(self.client, {"id": f"{context_id}"})
+        context.chat_history.append_user_message(prompt)
+
+        # Add a default LLM
+        context = with_llm(context=context, llm=OpenAI(client=self.client, model_name="gpt-4-0613"))
+
+        # AgentServices provide an emit function hook to access the output of running
+        # agents and tools. The emit functions fire at after the supplied agent emits
+        # a "FinishAction".
+        #
+        # Here, we show one way of accessing the output in a synchronous fashion. An
+        # alternative way would be to access the final Action in the `context.completed_steps`
+        # after the call to `run_agent()`.
+        output_blocks = []
+
+        def sync_emit(blocks: List[Block], meta: Metadata):
+            nonlocal output_blocks
+            output_blocks.extend(blocks)
+
+        context.emit_funcs.append(sync_emit)
+
+        # Get the agent
+        agent: Optional[Agent] = None
+        if hasattr(self, "agent"):
+            agent = self.agent
+        elif hasattr(self, "_agent"):
+            agent = self._agent
+
+        if not agent:
+            raise SteamshipError(
+                message="No Agent object found in the Agent Service. "
+                "Please name it either self.agent or self._agent."
+            )
+
+        self.run_agent(agent, context)
+
+        # Return the response as a set of multi-modal blocks.
+        return output_blocks
